requirements.txt
cloudflare==2.11.6
Django==4.2.2
django_recaptcha==3.0.0
djangorestframework==3.14.0
GoDaddyPy==2.3.4
json2html==1.3.0
pyOpenSSL==23.2.0
pyotp==2.8.0
pyTelegramBotAPI==4.12.0
Requests==2.31.0
schedule==1.2.0
